refactor: add new fouder mapper in infra customer
